Improve ECS and add entity class (#2)
This pull request includes improvements to the ECS system and adds a new entity class. The ECS system has been updated to support better entity management and component handling. Additionally, a new entity class has been added, which provides convenient methods for accessing and manipulating entity data. This will make it easier to work with entities in the game engine.